+++ src/Components/Dashbord/index.tsx
@@ -1,10 +1,13 @@
 import { Button, Space, Table } from 'antd';
 import { StyledLayout, StyledContent } from './styles';
 import Sidebar from 'Components/Sidebar';
-import { useAllNotApprovedRestDays , toApprovedOrDisapproveRestDay } from 'hooks/useUsers';
+import {
+  useAllNotApprovedRestDays,
+  toApprovedOrDisapproveRestDay,
+} from 'hooks/useUsers';
 import Loading from 'Components/Loading';
 import { lang } from 'language/en';
-import { APPROVED, DECLINED, CHANGED } from 'constants/statuses';
+import { APPROVED, DECLINED } from 'constants/statuses';
 import { IUserId } from 'hooks/types';
 
 const { Column, ColumnGroup } = Table;
@@ -14,7 +17,6 @@ const Dashbord = (): JSX.Element => {
 
   if (isLoading) return <Loading />;
   if (error instanceof Error) return <h1>Error: {error.message}</h1>;
-
 
   const putStatusApproved = (dataIndex: string, key: IUserId) => {
     toApprovedOrDisapproveRestDay({
@@ -30,13 +32,13 @@ const Dashbord = (): JSX.Element => {
       id: key.id,
       userId: dataIndex,
     });
-  }
+  };
 
   return (
     <StyledLayout>
       <Sidebar />
       <StyledContent>
-        <Table dataSource={data} pagination={{ pageSize: 10 }}  >
+        <Table dataSource={data} pagination={{ pageSize: 10 }}>
           <ColumnGroup title={lang.dashboard.userTitle}>
             <Column
               title={lang.dashboard.userFirstName}
@@ -59,11 +61,7 @@ const Dashbord = (): JSX.Element => {
             dataIndex="end_date"
             key="id"
           />
-          <Column
-            title={lang.dashboard.typeTitle}
-            dataIndex="type"
-            key="id" 
-          />
+          <Column title={lang.dashboard.typeTitle} dataIndex="type" key="id" />
           <Column
             title={lang.dashboard.actionTitle}
             dataIndex={['user', 'id']}
@@ -79,16 +77,13 @@ const Dashbord = (): JSX.Element => {
                   {lang.dashboard.approveButton}
                 </Button>
                 <Button
-                    htmlType="submit"
-                    type="link"
-                    onClick={() => putStatusDeclined(dataIndex, key)}
+                  htmlType="submit"
+                  type="link"
+                  onClick={() => putStatusDeclined(dataIndex, key)}
                 >
                   {lang.dashboard.declineButton}
                 </Button>
-                <Button
-                  htmlType="submit"
-                  type="link"
-                >
+                <Button htmlType="submit" type="link">
                   {lang.dashboard.editButton}
                 </Button>
               </Space>
